refactor(core): split candidate-construction modules and tests
Extract identity lineage buckets, represented-row filters, and pure
rank/logical helpers from the oversized identity-index and datum modules.
Keep buildPipelineCandidates as the sole outer seam.

Replace the 1549-line pipeline-build-candidates suite with focused
candidate-construction characterization files (mirroring the candidate-
store split), expand logical-value and frozen-bucket coverage, and move
collectColorChannelValues into its own test.
