+++ packages/core/tests/scale-color-collect.test.ts
@@ -1,0 +1,27 @@
+/**
+ * Color channel value collection for scale training.
+ */
+import { describe, expect, it } from "bun:test";
+import { fromAny } from "@total-typescript/shoehorn";
+
+describe("collectColorChannelValues", () => {
+  it("returns empty when no color mapping is present", async () => {
+    const { collectColorChannelValues } = await import("../src/pipeline/scale-color-collect.ts");
+    const table = fromAny({ has: () => false, discreteness: () => "continuous" });
+    const frames = fromAny([
+      {
+        binding: {
+          color: { field: null, scaledConstant: null },
+          fill: { field: null, scaledConstant: null },
+        },
+        colorValues: null,
+        fillValues: null,
+      },
+    ]);
+    expect(collectColorChannelValues("color", frames, table)).toEqual({
+      values: [],
+      anyDiscreteField: false,
+      anyField: false,
+    });
+  });
+});
